cleaned all warnings

+++ fifo.cpp
@@ -5,7 +5,7 @@ void Fifo::schedule(){
 		Request head = jobs.front();
 		if (head.size <= resources.size()){
 			std::queue<Request> _resources;
-			for (int i = 0; i < head.size; i++){
+			for (unsigned long i = 0; i < head.size; i++){
 				_resources.push(resources.front());
 				resources.pop();
 			}
